modify capistrano loading to automatically load V2 vs V3 capistrano tasks
