--- lib/sidekiq/capistrano3.rb
@@ -1,1 +1,0 @@
-load File.expand_path("../tasks/sidekiq.rake", __FILE__)
